requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pytesseract
Pillow